+++ tests/test_packages/test_cleveref.py
@@ -1,4 +1,75 @@
+import pytest
 from yalafi import parameters, parser, utils
+
+preamble = r'\usepackage[poorman]{cleveref}'
+
+
+def get_plain(latex):
+    parms = parameters.Parameters()
+    p = parser.Parser(parms)
+    plain, nums = utils.get_txt_pos(p.parse(preamble + latex))
+    assert len(plain) == len(nums)
+    return plain
+
+
+data_test_macros_latex = [
+
+    (r'A\crefname{type}{singular}{plural}B', 'AB'),
+    (r'A\Crefname{type}{singular}{plural}B', 'AB'),
+    (r'A\crefalias{counter}{type}B', 'AB'),
+
+]
+
+
+@pytest.mark.parametrize('latex,plain_expected', data_test_macros_latex)
+def test_macros_latex(latex, plain_expected):
+    plain = get_plain(latex)
+    assert plain == plain_expected
+
+
+preamble_sed = r'''\usepackage[poorman]{cleveref}
+\YYCleverefInput{tests/test_packages/cleveref.sed}'''
+
+
+def get_plain_sed(latex):
+    def read(file):
+        try:
+            with open(file) as f:
+                return True, f.read()
+        except:
+            return False, ''
+    parms = parameters.Parameters()
+    p = parser.Parser(parms, read_macros=read)
+    plain, nums = utils.get_txt_pos(p.parse(preamble_sed + latex))
+    assert len(plain) == len(nums)
+    return plain
+
+
+data_test_macros_sed = [
+
+    (r'\cref{a}', 'equation\N{NO-BREAK SPACE}(0)'),
+    (r'\Cref{a}', 'Equation\N{NO-BREAK SPACE}(0)'),
+    # (r'\crefrange{a}{b}', ''),
+    # (r'\Crefrange{a}{b}', ''),
+    (r'\cpageref{a}', 'page\N{NO-BREAK SPACE}0'),
+    (r'\Cpageref{a}', 'Page\N{NO-BREAK SPACE}0'),
+    (r'\cpagerefrange{a}{b}', 'pages\N{NO-BREAK SPACE}0 to\N{NO-BREAK SPACE}0'),
+    (r'\Cpagerefrange{a}{b}', 'Pages\N{NO-BREAK SPACE}0 to\N{NO-BREAK SPACE}0'),
+    (r'\namecref{a}', 'equation'),
+    (r'\nameCref{a}', 'Equation'),
+    (r'\namecrefs{a}', 'equations'),
+    (r'\nameCrefs{a}', 'Equations'),
+    (r'\lcnamecref{a}', 'equation'),
+    (r'\lcnamecrefs{a}', 'equations'),
+
+]
+
+
+@pytest.mark.parametrize('latex,plain_expected', data_test_macros_sed)
+def test_macros_sed(latex, plain_expected):
+    "Test different cleveref reference commands."
+    plain = get_plain_sed(latex)
+    assert plain == plain_expected
 
 
 latex_1 = r"""
@@ -17,6 +88,7 @@ eq.\N{NO-BREAK SPACE}(0)
 
 
 def test_1():
+    r"Test that `\newcommand{\YYCleverefInput}…` is ignored."
     def read(file):
         return True, sed_1
     parms = parameters.Parameters()
@@ -43,6 +115,7 @@ Equation\N{NO-BREAK SPACE}(0)
 
 
 def test_2():
+    "Test starred reference commands with hyperref."
     def read(file):
         return True, sed_2
     parms = parameters.Parameters()
@@ -92,6 +165,7 @@ stderr_4 = r"""*** LaTeX error: code in 't.tex', line 3, column 2:
 
 
 def test_4(capsys):
+    r"Test error creation for missing `\YYCleverefInput`."
     capsys.readouterr()
     parms = parameters.Parameters()
     p = parser.Parser(parms)
@@ -122,6 +196,7 @@ stderr_5 = r"""*** LaTeX error: code in 't.tex', line 4, column 2:
 
 
 def test_5(capsys):
+    "Test error creation for missing label in sed file."
     def read(file):
         return True, sed_5
     capsys.readouterr()
@@ -154,6 +229,7 @@ stderr_6 = r"""*** LaTeX error: code in 't.tex', line 2, column 2:
 
 
 def test_6(capsys):
+    "Test error creation for missing `poorman` option."
     def read(file):
         return True, sed_6
     capsys.readouterr()
@@ -181,6 +257,7 @@ eq.\N{NO-BREAK SPACE}(0)
 
 
 def test_7():
+    "Test special characters in equation label."
     def read(file):
         return True, sed_7
     parms = parameters.Parameters()
@@ -188,3 +265,36 @@ def test_7():
     toks = p.parse(latex_7)
     plain, pos = utils.get_txt_pos(toks)
     assert plain_7 == plain
+
+
+latex_8 = r"""
+\documentclass{book}
+\usepackage[german,main=british]{babel}
+\usepackage[poorman]{cleveref}
+\YYCleverefInput{latex.sed}
+\begin{document}
+    \begin{equation}
+        \label{eq1}
+        A=B
+    \end{equation}
+    \Cref{eq1}.
+\end{document}
+"""
+sed_8 = r"""
+5,6 s/\\Cref@equation@name /Equation/g 
+s/\\Cref{eq1}/\\Cref@equation@name \\nobreakspace \\textup {(\\ref {eq1})}/g 
+"""
+plain_8 = """
+      V-V-V\n    Equation\N{NO-BREAK SPACE}(0).
+"""
+
+
+def test_8():
+    "Test sed file for multi-language document."
+    def read(file):
+        return True, sed_8
+    parms = parameters.Parameters()
+    p = parser.Parser(parms, read_macros=read)
+    toks = p.parse(latex_8)
+    plain, pos = utils.get_txt_pos(toks)
+    assert plain_8 == plain